Change variable scope
"var" limits these variables to be local to this function.  Although declaring global vars is a have to and not a want to, in this case it is a requirement to be able to access the variables in other functions.  There is low risk of collision but I may consider renaming the variables later.

Working prototype!